fetch tickets

+++ script/LivrToken.s.sol
@@ -6,23 +6,22 @@ import "forge-std/Script.sol";
 import "../src/LivrToken.sol";
 
 contract LivrTokenScript is Script {
-  function setUp () public {}
+    function setUp() public {}
 
-  function run () public {
-    // Deploy LivrToken
-    // deploy("LivrToken", 100 * 10 ** 18, address(this));
-    uint privateKey = vm.envUint('PRIVATE_KEY');
-    address account = vm.addr(privateKey);
+    function run() public {
+        // Deploy LivrToken
+        // deploy("LivrToken", 100 * 10 ** 18, address(this));
+        uint256 privateKey = vm.envUint("PRIVATE_KEY");
+        address account = vm.addr(privateKey);
 
-    console.log("Account", account);
+        console.log("Account", account);
 
-    vm.startBroadcast(privateKey);
+        vm.startBroadcast(privateKey);
 
-    LivrToken livrToken = new LivrToken(100 * 10 ** 18, account);
+        LivrToken livrToken = new LivrToken(100 * 10 ** 18, account);
 
-    livrToken.MAX_SUPPLY();
+        livrToken.MAX_SUPPLY();
 
-    vm.stopBroadcast();
-
-  }
+        vm.stopBroadcast();
+    }
 }
